User Profile Update pt2
User signs in, can see their email, name, username, with a button to edit by all three of them. The password is obv not visible but you still have the option to edit them. The pages to edit each attribute are dynamically created (somewhat) and they properly update the db, too.