Updated ProfileController + profile.html
You can now see your current appointments. I want them to naturally go away on their own  after a certain amount of time but also add the ability to cancel if its within a certain time frame. Need to add cost variable to the rooms and add logic that calculates the price based on the amount of people and time selected.